C++ Insertion Sort: In Progress
I wanted to make a commit here to remember an important lesson. In c++ InsertionSort, I had an error on line 40. Whenever I ran the program, I would get an out of memory exception when the whileIndex was set to an U32. I figured out, through debugging, that this was caused the index being equal to 0, and then having 1 subtracted from it again. This caused it's value to be something 4 billion or 4 trillion. Some random, undefined value, that was different each time.

This happened becaused a U32 is Unsigned, so it can't become -1. The bug was fixed by changing the value to an S32, a Signed value, which can equal -1 without issues, and this allowed the code to execute as intended.

This is a win.